Split base_setup engine modules
