+++ cli/python/base_setup/checks.py
@@ -1,0 +1,40 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class ArtifactCheck:
+    name: str
+    ok: bool
+    message: str
+    fix: str
+    status: str = ""
+
+
+def check_to_json(check: ArtifactCheck) -> dict[str, str | bool]:
+    return {
+        "name": check.name,
+        "ok": check.ok,
+        "message": check.message,
+        "fix": check.fix,
+    }
+
+
+def check_to_doctor_json(check: ArtifactCheck) -> dict[str, str]:
+    return {
+        "status": doctor_status(check),
+        "name": check.name,
+        "message": check.message,
+        "fix": check.fix,
+    }
+
+
+def doctor_status(check: ArtifactCheck) -> str:
+    return check.status or ("ok" if check.ok else "error")
+
+
+def print_doctor_finding(status: str, name: str, message: str, fix: str = "") -> None:
+    print(f"{status:<5}  {name:<26}  {message}")
+    if fix:
+        print(f"       Fix: {fix}")
