+++ cli/python/base_setup/errors.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+
+class ArtifactError(RuntimeError):
+    pass
